requirements.txt
streamlit==1.18.1
pillow==9.3.0
pandas==1.4.4
numpy==1.23.5
scikit-learn==1.0.2
